+++ core/src/main/java/com/sonicgdx/sonicswirl/Sensor.java
@@ -41,7 +41,7 @@ public class Sensor {
 
         // An alternate expression to calculate block: ((chunkX * 128) + (tileX * 16) - xPosition));
 
-        byte height = TileMap.getTile(chunkX,chunkY,tileX,tileY).getHeight(block);
+        int height = TileMap.getTile(chunkX,chunkY,tileX,tileY).getHeight(block);
 
         float checkDistance = ((chunkY * 128) + (tileY * 16) + height) - yPosition;
 
@@ -105,7 +105,7 @@ public class Sensor {
         int block = Math.floorMod(MathUtils.round(yPosition),16); //Different behaviour for negative numbers compared to using %. For
         // example, -129 % 16 would return -1 which would cause an ArrayIndexOutOfBoundsException. Math.floorMod() would return a positive index in these cases.
 
-        byte width = TileMap.getTile(chunkX,chunkY,tileX,tileY).getWidth(block);
+        int width = TileMap.getTile(chunkX,chunkY,tileX,tileY).getWidth(block);
 
         //TODO change process if tile is flipped horizontally
         float checkDistance = ((chunkX * 128) + ((tileX + 1) * 16) - width) - xPosition;
